Now GUI shows stored contexts - not computing them
Computing the contexts can take a long time,
so it is not suitable to do this in the GUI.
Contexts are now computed with "python generate.py <schema_name>",
where <schema_name> is a schema defined in schemas.py.
Computed contexts are stored beneath the new "contexts" folder.
Most contexts are included in this commit, but a few
are still missing (for residuated lattices with n=11 or n=12 elements).