New Features: Context Sum, Context Aggregation
The GUI has two new buttons, "Sum" and "Aggregate".
Pressing "Sum" merges the contexts for lattices
with n elements, for all available n, into a single context.
This reproduces tables 10, 11 and 12 in Belohlavek and Vichodyl,
"Residuated Lattices of Size <= 12 -- Extended Version".
Pressing "Aggregate" counts the number of lattices with n elements
which have a certain attribute, for each n and each attribute,
and arranges the results in a table. This reproduces tables
7, 8 and 9 in the above mentioned paper.